requirements.txt
python-dateutil==2.6.1
Flask==0.12.2
flask-mongoengine==0.9.3
Flask-RESTful==0.3.6
mongoengine==0.14.3
WTForms==2.1
raven==6.2.1
blinker==1.4.0
bcrypt==3.1.3
PyJWT==1.5.3
flask_jwt_extended==3.3.1
requests_oauthlib==0.8.0
pycpfcnpj==1.2
Flask-Cors==3.0.3
django-dotenv==1.4.1
gunicorn==19.7.1
Flask-Caching==1.3.3
python-slugify==1.2.4
boto3==1.4.7
botocore==1.7.43
